pyproject.toml
[project]
name = "doodle-aux-mines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databases[aiosqlite]>=0.9.0",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.8.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]
